Move waitlist index init into waitlist db file
