+++ pkg/db/waitlist.go
@@ -7,10 +7,34 @@ import (
 	"github.com/SCE-Development/SCEvents/pkg/models"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
 )
 
 func GetWaitlistCollection() *mongo.Collection {
 	return Database().Collection("waitlists")
+}
+
+func InitWaitlistIndexes() error {
+	coll := GetWaitlistCollection()
+
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	_, err := coll.Indexes().CreateMany(ctx, []mongo.IndexModel{
+		{
+			Keys: bson.D{
+				{Key: "event_id", Value: 1},
+				{Key: "user_id", Value: 1},
+			},
+			Options: options.Index().SetUnique(true),
+		},
+		{
+			Keys: bson.D{
+				{Key: "event_id", Value: 1},
+			},
+		},
+	})
+	return err
 }
 
 func CreateWaitlistEntry(entry models.WaitlistEntry) error {
